docs!: remove help file
An odd commit, I guess. Keeping the help file up to date is awkward.
Previously I was using a workflow (removed in `f9071c8`) to auto
generate it. The major drawback (IMO) is polluting the commit history.
But the converter also had some issues (e.g, handling URLs and whatnot)

I can look into a more down to earth approach if this bothers someone.
But otherwise, the README will be the source of truth (and everything
else, it seems).