feat: improve multitab support
Previously, toggling ndv would be a noop if it was opened on another tab
(it would close that instance with no effect on the current tab). That
means you had to toggle twice for it to show up, which is lame.

To handle that, we can track the windows instances only for the current
tab. There are some downsides, which could be addressed with "proper"
multitab support (having ndv actually show up in multiples tabs at
once), but tracking state gets cumbersome.