feat!: do not open console on session initialization
The decision to open the terminal on session initialization was taken in
the project's early stages, with the rationale of being aligned with
nvim-dap's defaults. However, this is not really intuitive for some
users (https://github.com/igorlfs/nvim-dap-view/discussions/76). It also
doesn't align with the project's core principle of being unintrusive.
The sole reason it survived so long is to avoid a breaking change,
since, apparently, most users are not bothered by the current behavior.