perf(watches): massive rewrite to boost performance
Notable changes:
- No longer requires a listeners for the variables request
- No longer traverses the entire watch tree to fetch a specific variable
- When refreshing all expressions, wait for the last one to actually do
  the rendering

In addition to the performance improvements, the new implementation is
far cleaner. Variables have been renamed to be more consistent (although
that's sorta difficult when the object you're dealing with is called
"variable"). The flow for evaluation is much cleaner now, and follows a
similar structure between both requests.

Furthermore, some simplifications were made. Mostly thanks to the
refactoring, which lead to the discovery of some redundancies. What was
I thinking when I tried to have the same variable hold the response and
the err? Lolz