fix(console): set window options for the current buffer
Previously, console window options would always be set when showing the
console view, and they would also be applied for the "regular buffer"
(which would affect its scrolloff option).

Now these options are always set when a terminal is set as the current
buffer for the "regular window" and only for that buffer. Now the
"regular buffer" does not inherit these options (particularly, scrolloff
behaves as intended).

Because the options are now tied to the current buffer, we always have
to set them when updating the current term buf.