refactor: massively simplify tracking windows
Previously, when switching tabs, dap-view would keep track of its
windows for the new tab using filetypes. This is an extremely silly
idea. Couple of reasons why:
- It's awkward asking plugin authors to provide the filetype for their
  integrations. What if they may show multiple buffers within the same
  view? Chaos.
- Even if that wasn't a problem (so far it isn't), it's awkward
  nonetheless. In addition to checking the filetype, we have to track
  the `current_section` to try to be more certain that the window is the
  one we are looking for. Even still, it could lead to some false
  positives: what if the user had the REPL opened via `:DapToggleRepl`
  and also had dap-view on the same tab, showing the REPL? Once again,
  chaos.
With window local variables we can simply not bother. We don't even need
to use `vim.schedule` as a workaround for races.

For an even simpler approach, we could store the windows per tab (with
multiple there might be a better approach, though).