fix(listeners): delay setting up term buf until `configurationDone`
This partially reverts 395adc1

When `event_initialized` is triggered quickly, the term_buf might not be
defined for the session, leading to a race condition (see #125).

The original commit does not go in depth about any edge case that would
be covered by `event_initialized` (might have been a personal
preference, as I could be concerned about adapters not supporting
`configurationDone` - which is basically unheard of). Clearly, setting
up the terminal requires the session to be "fully configured".

Note: apparently, the workaround for
sessions view + js-debug-adapter + chrome + attach
still works with `configurationDone`