refactor: native sessions view
We no longer rely on nvim-dap's builtin sessions widget, we rolled our
own. Having full control of the rendering is nice, as it allows changing
stuff more easily on the future.

Kinda related to #124: there's an ongoing effort to also get rid of the
scopes widget. With that, nvim-dap-view would be completely independent
from the widgets, which I think is nice?

Having a single view based on widgets sucks, because it introduces some
abstractions and yada yada yada. It's simple enough to be implemented
"natively" with too much effort