fix(watches): ensure redraw always happens after user actions
This fixes some regressions from 0746f35, which ended up massively
complicating stuff and introducing a bunch of bugs.

As it turns out, dealing with the coroutine shenanigans is a lot easier
than anticipated: if we are wrapping with a coroutine, there's not much
to worry about. There's no need to handle call `yield` and `resume`
manually, as `nvim-dap` abstracts that.